Add DuckDuckGo Search feature
This change adds in a search feature using the duckduckgo instant
answer api [0]. While this API is a bit limited in scope it
should still provide a decent amount of relevant answers
when users query it. The current implementation is very
barebones and only handles a few cases. There is plenty
of room to improve the output of the results.

[0] https://duckduckgo.com/api